processBundleArtifacts(): work inside `bundle-artifacts-x86_64/`
The artifacts which this function processes actually come from the
`bundle-artifacts-x86_64` artifact. Let's let that be reflected in the
directory name.

This will also help in the upcoming support for embargoed releases,
where the files are no longer downloaded from workflow runs but will be
obtained via a different route.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>