self-hosted-runner: switch to federated credentials
Instead of storing stealable credentials in repository secrets, let's
create a managed identity instead and use federated credentials via
GitHub Actions' support for OpenID Connect.

This binds the authorization to GitHub workflows running in a specific
repository, and stealing the information won't enable an attacker to get
authorized to use the Azure subscription.

Note that this change _does_ require Client ID, Tenant ID and
Subscription ID to be stored separately as repository secrets (although
they do not _technically_ need to be kept secret, security is a game of
layers, so why give away this information?).

Also note that for some strange reason, the `contents: read` permission
seems to be lost when introducing a `permissions:` section. Therefore we
have to add it back explicitly, otherwise `actions/checkout` will fail
in a private repository.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>