self-hosted-runner: switch back to using `Azure/login`
We do not need the custom Action: `Azure/login` logs in using the Azure
CLI, and subsequent `az` calls work just fine.

So let's drop the complexity of the custom Action and go back to using
`Azure/login` instead.

The only downside is that we now need to specify the subscription ID
even though `az login` would work without it. But that's a small price
to pay, as the `delete-self-hosted-runner` workflow _still_ uses the
`Azure/login` Action and has to have that information as a repository
secret anyway.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>