self-hosted-runner: pass the post-deployment script without a URL
When running this workflow in a private repository, providing a public
URL to the post-deployment script simply would not work.

It is not even possible to use the `GITHUB_TOKEN` to construct an
`Invoke-WebRequest` call: The `GITHUB_TOKEN` lacks the permission to
access the resource.

So let's just pass this post-deployment script as a parameter.

Since it is somewhat large-ish, weighing 14kB, let's compress it. And
since the compressed file is binary and cannot easily be passed around,
let's Base64-encode it. The result is still somewhat large (5.6kB) but
at least this works and still leaves some room for additional stuff to
be put into the post-deployment script.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>