Add a new workflow to build embargoed Pacman packages
Embargoed releases come in various shapes and forms. One of the more
complex shape involves integrating embargoed versions of components such
as Git Credential Manager or Git LFS.

To support this, let's add a workflow (copy-edited from
`build-and-deploy` with substantial changes, so substantial as to
justify a separate workflow) that builds a Pacman package.

Since both input and output artifacts are under strict embargo, let's
use GitHub Releases in the embargoed org for that: A GitHub Release to
host the source files (such as, say, `git-lfs-windows-386-v3.6.1.zip`),
and a GitHub Release to gather the created package files.

Since multiple runs may be required (for packages that cannot be built
within the same Git SDK for all supported CPU architectures), this new
workflow creates or reuses a GitHub Release that tags the revision from
which the package is built, using `<package-name>-<pkgver>-<pkgrel>` as
tag name.

The idea is that the output of this workflow can be the input to an
upcoming `install-embargoed-package` workflow that installs the
respective packages into the respective (embargoed) Git SDKs.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>