prepare-embargoed-package: reuse Release if it exists already
When building, say, MSYS packages for both i686 and x86_64, it will
require separate workflow runs per architecture. Yet the resulting
packages should go into the same GitHub Release. So let's be prepared
for that situation.

Note: Technically it is possible that concurrent workflow runs try to
create a Release at the very same moment and that GitHub is a bit
overwhelmed by that and creates none. This is highly theoretical,
though, so let's not worry about it, and in the unlikely event that it
does happen, manually re-run the failed runs individually, thereby
preventing the race altogether.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>