github-release: add a function to download assets from a GitHub Release
It requires an amazingly unintuitive process to do something as
commonplace as downloading release assets. That's why I originally
preferred to let the GitHub CLI deal with it an be done with it.

However, I want to avoid the necessity to have that GitHub CLI installed
on self-hosted runners, so let's reimplement this using the plain (or
actually, not so plain, as indicated above) REST API.

Note that I still punt on replacing the `curl` call; It strikes me as
not so bad, given that a `curl` version that seems good enough to handle
my request is installed on all-but-ancient Windows versions by default
(see https://curl.se/windows/microsoft.html for details).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>