build-embargoed-package: avoid relying on GitHub CLI
While convenient, it is a cumbersome dependency to have; Not even GitHub
Actions' runners come with a copy of it.

So let's bite the bullet and replace the call to `gh release download`
with some JavaScript (which is eminently more convenient to execute,
given that the `actions/github-script` Action takes care of instaling
node.js).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>