release-git: require the announcement email to be sent manually
For quite a while now, those emails were rejected by the Git mailing
list due to some Spamhaus rule. This is quite annoying, but cannot be
helped because, well, because reasons.

So let's just print out the announcement email to the log, thusly
requiring a manual step after the release is done.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>