Add a GitHub workflow to deploy embargoed Pacman packages
Typically weeks (or in the case of v2.47.2, over two months as per
request of the then-interim Git maintainer) go by between building those
packages and actually deploying them to Git for Windows' Pacman
repository.

To help with the deployment (and make it as boring and uneventful as
possible, see https://zachholman.com/posts/deploying-software), let's
add some automation.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>