git-artifacts(snapshots): work around a bug in `git fetch <bundle>`
As part of the snapshot builds via the `git-artifacts` workflow, we
try to import the tag from a bundle.

This uncovered a bug that is over 17 years old, and which I am trying
to fix via https://github.com/gitgitgadget/git/pull/1857.

The symptom is that just after fetching the refs, when looking through
the fetched revisions whether a recursive fetch is needed, Git fails to
`mmap()` the newly-imported packfile. The symptom looks like this:

  From [...]/git.bundle
   * tag                     <tag> -> FETCH_HEAD
   * [new tag]               <tag> -> <tag>
  fatal: mmap: could not determine filesize

Curiously, this only happens on 32-bit Windows, not on 64-bit Windows,
nor on Linux or macOS. The explanation is most likely to be found in
how quickly file descriptor values are used again after closing them,
which would appear to be _really_ quickly on i686 Windows.

Be that as it may, to work around this issue, we simply avoid any
operation that would need to access the just-imported packfile in
`git fetch <bundle>`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>