build-and-deploy/release-git: support the new Pacman repository
In https://github.com/git-for-windows/build-extra/pull/595, I started
the official transition of Git for Windows' Pacman repository away from
an Azure Blob container in my personal account to
https://github.com/git-for-windows/pacman-repo.

For the transition period, this still requires the token to upload to
the Azure Blob container, but it also already requires the token to push
to `pacman-repo` and to create GitHub releases there.

The two users of `pacman-helper.sh quick_add` that now need this token
are hereby adapted accordingly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>