git-artifacts: _really_ stop building i686 artifacts except mingit
As announced multiple times and also documented in detail at
https://gitforwindows.org/32-bit, Git for Windows will no longer
publish 32-bit installers, portable Gits and such.

The only i686 artifacts that will be published are the MinGit ones,
until April 2029.

In d308f34 (git-artifacts: stop building i686 artifacts except mingit,
2025-02-25), I already tried to address this. But I forgot that the
default, if left unspecified which artifacts to build, is to build all
of them (with "all" being hard-coded also in the `Configure environment`
step of the `git-artifacts` workflow).

So let's adjust this step, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>