release-git(announcement): add a helpful comment
Since the Git mailing list so helpfully denies us the ability to post
emails from Azure IPs, Git for Windows' announcement emails have to be
sent manually again, which is a hassle.

To make the process easier, add a PR comment with a direct link to the
artifact containing the announcement email, and while at it also prepare
an announcement to share on Bluesky.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>