release(skeet): avoid saying "release candidate"
There is nothing wrong with release candidates, they are great: Once
enough people have tested a release candidate, it can be confidently
switched to a full release.

So what is wrong about calling Git for Windows' -rc versions (or for
that matter, Git's -rc versions "release candidates" is that, well,
they are _not at all candidates_ to be turned into full releases.

So let's call them what they are: previews.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>