git-artifacts: validate the installers as part of the workflow
We already do that in the `git-artifacts` workflows in the `git-sdk-*`
repositories, so it is only logical to do the same in the
`git-artifacts` workflow in `git-for-windows-automation`.

Note: I copy/edited the version from the `git-sdk-arm64` repository's
`git-artifacts` workflow because that is quite a bit ahead of the other
`git-sdk-*` repositories' versions.

There is one notable deviation from that workflow: The `installer.log`
file is no longer shown in the workflow run's log. The reason is: it is
lengthy, and the grouping does not work due to GitHub's web UI
implementation. GitHub's web UI insists on loading only a small chunk of
the logs at a time, missing the end of the group, and simply not showing
the relevant part (which comes after the log). The `installer.log` file
is uploaded as an artifact anyway, so not much is lost.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>