git-artifacts: suggest to validate installer in a PR comment
It is currently a bit cumbersome to find the download link to the
installer artifact, and I even missed that it was available for testing
already while the `aarch64` stuff was still being built.

Let's imitate c99c7f2 (release-git(announcement): add a helpful comment,
2025-03-11) and add a helpful PR comment not only when the release
finished, but also when the x86_64 installer was successfully uploaded
and is ready for manual testing.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>