vscode: add a Debug launch configuration
This repository has quite a few node.js scripts in it, and sometimes
it is useful to debug them using a proper debugger, using
single-stepping, break points, watch expressions, just like the one that
comes with VS Code.

So let's add a launch configuration that allows to do just that. By
default, it starts the current file. That is not generally useful
because these node.js scripts are meant to be included in the GitHub
workflows and the function are expected to be called from
`actions/github-script` steps. However, by appending something like
this, debugging now gets a lot easier:

  ;(async () => {
    console.log('Hello world!')
  })().catch(e => {
    console.error(e)
    process.exit(1)
  })

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>