tag-git: accommodate for upstream's `.txt` -> `.adoc` rename
This is the gazillionth follow-up to that rename; It was noticed in
https://github.com/git-for-windows/build-extra/pull/607 that this needed
yet another adjustement, this time in Git for Windows' release notes. In
that PR, the `please.sh` script was adjusted, which indeed used to be
how the release notes generation was automated.

In the meantime, we have moved from an Azure Pipelines-based
automation to a GitHub Actions-based automation, and the `please.sh`
script is no longer used to generate those; Instead the
`update-scripts/tag-git.sh` script in this here repository is used for
that purpose.

Therefore, the latter script needs to be adjusted, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>