build-and-deploy: drop work around for cURL v8.11.0
When deploying curl-8.11.0-2, a problem needed to be worked around:
that cURL version, v8.11.0, had a bug that needed to be fixed, but the
fix could not be deployed because of that bug. To work around this
problem, the `build-and-deploy` workflow needed to force-downgrade
the `curl` package.

This workaround is no longer needed, therefore we can drop it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>