build-and-deploy: set MINGW_ARCH appropriately for i686/x86_64 builds
In the future, we want to build the packages in the appropriate git-sdk
for the given CPU architecture. Right now, we still build i686/x86_64
MINGW packages in git-sdk-64.

Let's respect the `ARCHITECTURE` environment variable (which is left
empty in the current workflow runs that want to build i686/x86_64 MINGW
packages), and set `MINGW_ARCH` appropriately when the `ARCHITECTURE`
is set to `i686` or `x86_64` (which we will do in the future).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>