ensure-not-yet-deployed: use the new Pacman repository's URL
A while ago I started the migration from the old Pacman repository
(which was hosted in my personal Azure account) to the new one (which is
hosted in a Git repository on GitHub).

For the transition period, it was okay to use the old URL to see whether
a package is already deployed or not, as the package files are deployed
to both old and new Pacman repository. However, in preparation for
retiring the old Pacman repository, we need to use the new URL
consistently throughout Git for Windows' automation.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>