tag-git: work around `.txt`/`.adoc` inconsistency
One of the things in the Git project that you either grow to love or
instead to wish away is the somewhat consistent tendency to be
inconsistent.

This commit concerns itself with the fact that the release notes of
v2.49.1 are stored in a `.txt` file in `Documentation/RelNotes/`. All
other files in that directory have the `.adoc` extension ever since that
ill-executed rename of 1f010d6bdf (doc: use .adoc extension for AsciiDoc
files, 2025-01-20). Funnily enough, this is only true for the revision
tagged as v2.49.1, but the one tagged as v2.50.1 has them renamed to
`.adoc` files.

I'll just do what I always do: Deal with the fall-out in a pragmatic
way, as there isn't really any good alternative to that approach.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>