pushRepositoryUpdate: allow reusing local clones
When publishing a previously-embargoed release, it is often the case
that there are _multiple_ releases to be published.

When doing this manually, there is no clean slate (as would there be
when running the script in a GitHub workflow), and therefore, say, the
`build-extra` clone might exist locally already.

Rather than deleting that clone and re-clone (which is a waste of
bandwidth), let's just reuse the local clone, ensuring that the current
branch is updated to the same state as it would be in case of a fresh
clone.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>