deploy-embargoed-package: mirror the check-run to the originating repository
With this change, the workflow will mirror a Check Run to the commit
from which the GitHub Release with the embargoed packages was created.

This is particularly useful when publishing embargoed `mingw-w64-git`
packages, as those Check Runs will be visible in the PR where the
embargoed packages were created, and thus provide a convenient way to
track the progress of the embargoed builds.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>