init-g4w-sdk-for-pacman: fix caching
actions/cache/restore does a prefix match on the cache key to
restore. [1] This can lead to restoring a different cache than we expect.
In that case the cache-hit output is set to false, which further
disturbs our workflows, since a cache was restored, but the output
claims a cache miss.

To work around this, turn the makepkg suffix into an infix.

[1] https://github.com/actions/cache/issues/1433#issuecomment-2259294936

Signed-off-by: Matthias Aßhauer <mha1993@live.de>