github-release: prepare for immutable releases
There is a new initiative to increase confidence in GitHub Releases:
https://github.blog/changelog/2025-10-28-immutable-releases-are-now-generally-available/

The official guidance is to create a draft release, then attach all assets, then publish:
https://docs.github.com/en/code-security/supply-chain-security/understanding-your-software-supply-chain/immutable-releases#best-practices-for-publishing-immutable-releases

We _almost_ do that already. The only exception is the additional update
to create a discussion for each new release. This was separated out
earlier in response to a missing permission of the GitHub App used to
perform the job, but is likely incompatible with immutable releases.

So let's revert b59f512 (github-release: create the Discussion
separately, 2024-09-18).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>