git-artifacts: prepare for the big split of mingw-w64-git
In https://github.com/git-for-windows/MINGW-packages/pull/172, the
`mingw-w64-git` package will be split into multiple packages, with the
new primary package as Git for Windows is concerned being the
`mingw-w64-git-for-windows-addons` one.

Prepare for that future by simply adding (almost) all of the packages
that were built to the command-line via `--pkg` options.

While at it, simplify the way the version option passed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>